sextougit add .git add .

+++ src/main/java/com/example/demo/service/CompanyService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 
+import com.example.demo.excepetions.RecursoNaoEncontradoException;
 import com.example.demo.model.Company;
 import com.example.demo.repository.CompanyRepository;;
 
@@ -21,8 +22,9 @@ public class CompanyService {
         return companyRepository.findAll();
     }
 
-    public Optional<Company> buscarPorId(Long id) {
-        return companyRepository.findById(id);
+    public Company buscarPorId(Long id) {
+        return companyRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Company com ID "+id+" não encontrado"));
     }
 
     public Company salvarCompany(Company company) {
@@ -30,6 +32,11 @@ public class CompanyService {
     }
 
     public void deletarCompany(Long id) {
+
+        if (!companyRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Company com ID "+id+" não encontrado");
+        }
+
         companyRepository.deleteById(id);
     }
 }
